Task 3 and almost Task 2
Task 2 lost the rotation on movement using A and D, I'll look into it.
On Task 3, items disappear, start re-appearing slowly changing the transparency and once they are 100% spawned again, then the player can collect them. They can be collected when they are rotating, which I decided to add to improve readability.