Final Push
Bibliography:
Video I used on how to enable and disable box collider for the food items: https://youtu.be/Bl9SE73xqI4

Used the Unity Documentation website for additional help:
https://www.bing.com/ck/a?!&&p=ddea8692876617e6JmltdHM9MTcxNzcxODQwMCZpZ3VpZD0wOTliNWMxMi1jNWEzLTY3YmQtMDk5MC00ODExYzQ0ZTY2MTImaW5zaWQ9NTIxMw&ptn=3&ver=2&hsh=3&fclid=099b5c12-c5a3-67bd-0990-4811c44e6612&psq=unity+documentation&u=a1aHR0cHM6Ly9kb2NzLnVuaXR5LmNvbS8&ntb=1

Nathan has helped me a great deal with the strafing animations and it's implementation in the player controller script. Enzo has also helped me with creating a list for all the food items, so that a random food will appear when one is collected.