Final changes in animations
Strafing is now working as intended after some re-imports of animations and changes to code logic to make strafing to FALSE so it would exit its state. I had a lot of help from Musse to make this work, but I did the rest on my own.
I did got recommended a video from Chanakan for the camera task and it worked for me (https://www.youtube.com/watch?v=LbDQHv9z-F0)